Split messaging callbacks into new file: serialdata.messaging.c
